OAK-11766 Write Throttling Mechanism - Session.save() delay
